requirements.txt
amqp==5.0.9
asgiref==3.4.1
atomicwrites==1.4.0
attrs==19.3.0
billiard==3.6.4.0
celery==5.2.3
certifi==2020.4.5.1
cffi==1.15.0
chardet==3.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.0.5
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-extra-fields==3.1.1
flower==1.0.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.0.4
humanize==3.13.1
idna==2.9
importlib-metadata==1.6.0
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
more-itertools==8.2.0
oauthlib==3.1.1
packaging==21.3
Pillow==8.3.1
pluggy==0.13.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psycopg2-binary==2.8.6
py==1.8.1
pycparser==2.21
PyJWT==2.0.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
python-dotenv==0.18.0
python3-openid==3.2.0
pytz==2021.3
redis==4.1.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
tornado==6.1
uritemplate==4.1.1
urllib3==1.25.9
vine==5.0.0
wcwidth==0.1.9
wrapt==1.13.3
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.4.0
